requirements.txt
black==23.12.1
canvasapi==3.2.0
django-cors-headers==4.3.1
django-filter==23.5
Django==5.0
djangorestframework==3.14.0
Faker==24.2.0
gunicorn==23.0.0
Markdown==3.5.2
psycopg2==2.9.9
pyjwt==2.8.0
python-dotenv==1.0.1